+++ tests/include/policies_variables_and_functions.h
@@ -14,7 +14,7 @@ using namespace sam;
 
 struct PoliciesVariablesAndFunctions {
   
-  std::vector<double> values {
+  std::vector<float> values {
     0, -0, .0, -.0, 0.0, -0.0,
     .1, -.1, 0.1, -0.1,
     1, -1
